% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolstats-package.R
\docType{package}
\name{mongolstats-package}
\alias{mongolstats}
\alias{mongolstats-package}
\title{mongolstats: Tidy Client for the Mongolian NSO PXWeb Statistics API}
\description{
A tidyverse-friendly client for the National Statistics Office of Mongolia
PXWeb API (data.1212.mn) with helpers to discover tables, variables, and
periods, and to fetch statistical data. For maps, join the results to the
boundaries in the companion package mongolmaps by NSO code.
}
\section{Main Functions}{

\describe{
\item{nso_search}{Find tables by keyword}
\item{nso_tables}{List all tables}
\item{nso_dims, nso_dim_values}{Inspect a table's dimensions and values}
\item{nso_table_periods, nso_latest_periods}{Periods available in a table}
\item{nso_data}{Fetch statistical data from a table}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://temuulene.github.io/mongolstats/}
  \item \url{https://data.1212.mn/pxweb/}
  \item Report bugs at \url{https://github.com/temuulene/mongolstats/issues}
}

}
\author{
\strong{Maintainer}: Temuulen Enebish \email{temuulen@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Temuulen Enebish \email{temuulen@gmail.com} [copyright holder]
}

}
\keyword{internal}
