% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nso_cache_status}
\alias{nso_cache_status}
\title{Cache status}
\usage{
nso_cache_status()
}
\value{
A list with \code{enabled}, \code{dir}, and \code{has_cache}.
}
\description{
Report current cache configuration and basic stats.
}
\examples{
nso_cache_status()
}
