% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itms.R
\name{nso_itms}
\alias{nso_itms}
\alias{nso_itms_detail}
\alias{nso_itms_search}
\title{Superseded discovery functions}
\usage{
nso_itms()

nso_itms_detail(tbl_id)

nso_itms_search(query, fields = c("tbl_eng_nm", "tbl_nm"))
}
\arguments{
\item{tbl_id}{Table identifier.}

\item{query}{A single keyword, matched literally and case-insensitively.}

\item{fields}{Columns of \code{\link[=nso_tables]{nso_tables()}} to search.}
}
\value{
As the newer functions.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

These are the original names of three discovery functions. They keep
working, but new code should use the newer names:
\itemize{
\item \code{nso_itms()} is \code{\link[=nso_tables]{nso_tables()}}.
\item \code{nso_itms_detail()} is \code{\link[=nso_variables]{nso_variables()}}.
\item \code{nso_itms_search(query)} is \code{\link[=nso_search]{nso_search(query, fixed = TRUE)}}.
}
}
\keyword{internal}
