% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{as_px_query}
\alias{as_px_query}
\title{Convert a query to a PXWeb body}
\usage{
as_px_query(x, lang = .px_lang())
}
\arguments{
\item{x}{An \code{nso_query} object.}

\item{lang}{PX language: "en" or "mn" (defaults to current option).}
}
\value{
A list suitable to send as JSON body to PXWeb.
}
\description{
Convert a query to a PXWeb body
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
q <- nso_query("DT_NSO_0300_001V2", list(Year = "2023"))
body <- as_px_query(q)
\dontshow{\}) # examplesIf}
}
