% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{mn_boundaries_normalize}
\alias{mn_boundaries_normalize}
\title{Add normalized name columns to boundaries}
\usage{
mn_boundaries_normalize(g, name_col = "shapeName")
}
\arguments{
\item{g}{sf object from \code{mn_boundaries()}}

\item{name_col}{Column with English names (default 'shapeName').}
}
\value{
sf with \code{name_std} column added.
}
\description{
Add normalized name columns to boundaries
}
