% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{mn_boundary_keys}
\alias{mn_boundary_keys}
\title{Boundary keys/crosswalk helper}
\usage{
mn_boundary_keys(level = "ADM1")
}
\arguments{
\item{level}{Boundary level.}
}
\value{
tibble with key columns from GeoBoundaries and normalized names.
}
\description{
Boundary keys/crosswalk helper
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
keys <- mn_boundary_keys("ADM1")
head(keys)
\dontshow{\}) # examplesIf}
}
