% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{nso_cache_enable}
\alias{nso_cache_enable}
\title{Enable or configure caching}
\usage{
nso_cache_enable(dir = NULL)
}
\arguments{
\item{dir}{Directory for cache; defaults to user cache dir.}
}
\value{
Cache directory path.
}
\description{
Caches table lists and codebooks on disk to speed up repeated calls.
}
