% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itms.R
\name{nso_itms_by_sector}
\alias{nso_itms_by_sector}
\title{List tables under a sector or sub-sector (PXWeb path)}
\usage{
nso_itms_by_sector(list_id)
}
\arguments{
\item{list_id}{Path string from \code{nso_sectors()}/\code{nso_subsectors()} \code{id}.}
}
\description{
List tables under a sector or sub-sector (PXWeb path)
}
