% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itms.R
\name{nso_itms_detail}
\alias{nso_itms_detail}
\alias{nso_variables}
\title{Get variable codes for a table (PXWeb)}
\usage{
nso_itms_detail(tbl_id)

nso_variables(tbl_id)
}
\arguments{
\item{tbl_id}{Table identifier (e.g., "DT_NSO_0300_001V2").}
}
\description{
Get variable codes for a table (PXWeb)
}
