% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_index.R
\name{nso_rebuild_px_index}
\alias{nso_rebuild_px_index}
\title{Rebuild PXWeb index and optionally write to a file}
\usage{
nso_rebuild_px_index(
  path = file.path("inst", "extdata", "px_index.json"),
  write = TRUE
)
}
\arguments{
\item{path}{Output path for JSON (default 'inst/extdata/px_index.json' in dev trees)}

\item{write}{Whether to write JSON to \code{path} (default TRUE)}
}
\value{
tibble index
}
\description{
Rebuild PXWeb index and optionally write to a file
}
