% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{nso_search}
\alias{nso_search}
\title{Search NSO tables}
\usage{
nso_search(query, sector = NULL, fields = c("tbl_eng_nm", "tbl_nm"))
}
\arguments{
\item{query}{Search string (regex, case-insensitive).}

\item{sector}{Optional sector/subsector \code{list_id} to filter results.}

\item{fields}{Character vector of fields to search within.}
}
\value{
Tibble of matching tables.
}
\description{
Search NSO tables
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
nso_search("population")
\dontshow{\}) # examplesIf}
}
