% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{nso_sectors}
\alias{nso_sectors}
\title{List top-level categories (PXWeb NSO root)}
\usage{
nso_sectors()
}
\value{
tibble with \code{id}, \code{type}, \code{text}
}
\description{
List top-level categories (PXWeb NSO root)
}
