% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{nso_subsectors}
\alias{nso_subsectors}
\title{List children for a given path (PXWeb)}
\usage{
nso_subsectors(subid)
}
\arguments{
\item{subid}{Path id from \code{nso_sectors()}/\code{nso_subsectors()} (e.g., 'Population, household' or 'Population, household/1_Population, household')}
}
\description{
List children for a given path (PXWeb)
}
