% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{nso_subsectors}
\alias{nso_subsectors}
\title{List children for a given path (PXWeb)}
\usage{
nso_subsectors(subid)
}
\arguments{
\item{subid}{Path id from \code{nso_sectors()}/\code{nso_subsectors()}
(e.g., 'Population, household' or 'Population, household/1_Population, household')}
}
\value{
A tibble with columns: \code{id}, \code{type}, \code{text}.
}
\description{
List children for a given path (PXWeb)
}
\examples{
\dontshow{if (curl::has_internet()) withAutoprint(\{ # examplesIf}
sectors <- nso_sectors()
nso_subsectors(sectors$id[1])
\dontshow{\}) # examplesIf}
}
